requirements.txt
llama_cpp_python==0.3.9
gradio==5.29.0
setuptools==63.2.0